Modify iRacing Dollars & Credits to Accept Decimal

+++ src/Aydsko.iRacingData/Member/Account.cs
@@ -6,10 +6,10 @@ namespace Aydsko.iRacingData.Member;
 public class Account
 {
     [JsonPropertyName("ir_dollars")]
-    public int IRacingDollars { get; set; }
+    public decimal IRacingDollars { get; set; }
 
     [JsonPropertyName("ir_credits")]
-    public int IRacingCredits { get; set; }
+    public decimal IRacingCredits { get; set; }
 
     [JsonPropertyName("status")]
     public string Status { get; set; } = default!;
